Change message size

+++ hellfireos/app/noc_img_filter/process_info.c
@@ -1,12 +1,12 @@
 #include <hellfire.h>
 #include "process_info.h"
 
-uint16_t * get_content(uint16_t * message)
+int8_t * get_content(int8_t * message)
 {
         return &message[1];
 }
 
-uint8_t * get_process_type(uint16_t * message)
+int8_t * get_process_type(int8_t * message)
 {
         return message[0];
 }
